+++ fem/fe/fe_base.cpp
@@ -12,6 +12,7 @@
 // Finite Element Base classes
 
 #include "fe_base.hpp"
+#include "face_map_utils.hpp"
 #include "../coefficient.hpp"
 
 namespace mfem
@@ -368,6 +369,12 @@ const DofToQuad &FiniteElement::GetDofToQuad(const IntegrationRule &,
 {
    MFEM_ABORT("method is not implemented for this element");
    return *dof2quad_array[0]; // suppress a warning
+}
+
+void FiniteElement::GetFaceMap(const int face_id,
+                               Array<int> &face_map) const
+{
+   MFEM_ABORT("method is not implemented for this element");
 }
 
 FiniteElement::~FiniteElement()
@@ -2425,6 +2432,55 @@ void NodalTensorFiniteElement::SetMapType(const int map_type)
    {
       basis1d.ScaleIntegrated(map_type == VALUE);
    }
+}
+
+void NodalTensorFiniteElement::GetFaceMap(const int face_id,
+                                          Array<int> &face_map) const
+{
+   const int dof1d = order + 1;
+   int n_face_dofs = pow(dof1d, dim - 1);
+   std::vector<int> offsets, strides;
+   switch (dim)
+   {
+      case 1:
+         offsets = {(face_id == 0) ? 0 : dof1d - 1};
+         break;
+      case 2:
+         strides = {(face_id == 0 || face_id == 2) ? 1 : dof1d};
+         switch (face_id)
+         {
+            case 0: offsets = {0}; break; // y = 0
+            case 1: offsets = {dof1d - 1}; break; // x = 1
+            case 2: offsets = {(dof1d-1)*dof1d}; break; // y = 1
+            case 3: offsets = {0}; break; // x = 0
+         }
+         break;
+      case 3:
+      {
+         const auto f = GetFaceNormal3D(face_id);
+         const int face_normal = f.first, level = f.second;
+         if (face_normal == 0) // x-normal
+         {
+            offsets = {level ? dof1d-1 : 0};
+            strides = {dof1d, dof1d*dof1d};
+         }
+         else if (face_normal == 1) // y-normal
+         {
+            offsets = {level ? (dof1d-1)*dof1d : 0};
+            strides = {1, dof1d*dof1d};
+         }
+         else if (face_normal == 2) // z-normal
+         {
+            offsets = {level ? (dof1d-1)*dof1d*dof1d : 0};
+            strides = {1, dof1d};
+         }
+         break;
+      }
+   }
+
+   // same number of DOFs in each dimension, repeat dof1d (dim - 1) times
+   std::vector<int> n_dofs(dim - 1, dof1d);
+   FillFaceMap(n_face_dofs, offsets, strides, n_dofs, face_map);
 }
 
 VectorTensorFiniteElement::VectorTensorFiniteElement(const int dims,
